HTML-806: Problem with Viewing/Editing Obs Groups that have multiple checkboxes with the same question
