+++ api/src/test/resources/org/openmrs/module/htmlformentry/include/obsGroupWithMultipleCheckboxAnswersForSameQuestionConcept.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<htmlform>
+    Date: <encounterDate/>
+    Location: <encounterLocation/>
+    Provider: <encounterProvider
+        role="Provider"/>
+
+        <obsgroup groupingConceptId="7">
+            <obs conceptId="1000" answerConceptId="1001" answerLabel="foo1" labelText=""/>
+            <obs conceptId="1000" answerConceptId="1002" answerLabel="foo2" labelText=""/>
+            <obs conceptId="1000" answerConceptId="1003" answerLabel="foo3" labelText=""/>
+        </obsgroup>
+
+</htmlform>
